+++ test/foundry/mocks/MockYnETHERC4626.sol
@@ -36,20 +36,20 @@ contract MockYnETHERC4626 is IynETH, AccessControlUpgradeable, ERC4626Upgradeabl
     error MinimumStakeBoundNotSatisfied();
     error StakeBelowMinimumynETHAmount(uint256 ynETHAmount, uint256 expectedMinimum);
     error Paused();
-    error ValueOutOfBounds(uint value);
+    error ValueOutOfBounds(uint256 value);
 
 
     IStakingNodesManager public stakingNodesManager;
     IRewardsDistributor public rewardsDistributor;
-    uint public allocatedETHForDeposits;
+    uint256 public allocatedETHForDeposits;
     bool public depositsPaused;
     // Storage variables
 
 
     /// @dev The value is in basis points (1/10000).
-    uint public exchangeAdjustmentRate;
+    uint256 public exchangeAdjustmentRate;
 
-    uint public totalDepositedInPool;
+    uint256 public totalDepositedInPool;
 
     struct ReInit {
 
@@ -64,7 +64,7 @@ contract MockYnETHERC4626 is IynETH, AccessControlUpgradeable, ERC4626Upgradeabl
          
     }
 
-    function deposit(uint assets, address receiver) public override returns (uint shares) {
+    function deposit(uint256 assets, address receiver) public override returns (uint256 shares) {
 
         if (depositsPaused) {
             console.log("System is paused");
@@ -114,8 +114,8 @@ contract MockYnETHERC4626 is IynETH, AccessControlUpgradeable, ERC4626Upgradeabl
         return _convertToShares(assets, Math.Rounding.Floor);
     }
 
-    function totalAssets() public view override returns (uint) {
-        uint total = 0;
+    function totalAssets() public view override returns (uint256) {
+        uint256 total = 0;
         // allocated ETH for deposits pending to be processed
         total += totalDepositedInPool;
         /// The total ETH sent to the beacon chain 
@@ -123,10 +123,10 @@ contract MockYnETHERC4626 is IynETH, AccessControlUpgradeable, ERC4626Upgradeabl
         return total;
     }
 
-    function totalDepositedInValidators() internal view returns (uint) {
+    function totalDepositedInValidators() internal view returns (uint256) {
         IStakingNode[]  memory nodes = stakingNodesManager.getAllNodes();
-        uint totalDeposited = 0;
-        for (uint i = 0; i < nodes.length; i++) {
+        uint256 totalDeposited = 0;
+        for (uint256 i = 0; i < nodes.length; i++) {
             totalDeposited += nodes[i].getETHBalance();
         }
         return totalDeposited;
@@ -145,7 +145,7 @@ contract MockYnETHERC4626 is IynETH, AccessControlUpgradeable, ERC4626Upgradeabl
         totalDepositedInPool += msg.value;
     }
 
-    function withdrawETH(uint ethAmount) public onlyStakingNodesManager override {
+    function withdrawETH(uint256 ethAmount) public onlyStakingNodesManager override {
         require(totalDepositedInPool >= ethAmount, "Insufficient balance");
 
         payable(address(stakingNodesManager)).transfer(ethAmount);
